1207 カレンダーの一括削除などを行うとエラーが発生する不具合の修正

+++ modules/Calendar/actions/MassDelete.php
@@ -14,15 +14,35 @@ class Calendar_MassDelete_Action extends Vtiger_MassDelete_Action {
 		$adb = PearDatabase::getInstance();
 		$moduleName = $request->getModule();
 		$moduleModel = Vtiger_Module_Model::getInstance($moduleName);
+		$cvId = $request->get('viewname');
 
 		if ($request->get('selected_ids') == 'all' && $request->get('mode') == 'FindDuplicates') {
             $recordIds = Vtiger_FindDuplicate_Model::getMassDeleteRecords($request);
         } else {
             $recordIds = $this->getRecordsListFromRequest($request);
         }
-		$cvId = $request->get('viewname');
+		
+		$skipRecords = [];
+		asort($recordIds, SORT_NUMERIC);
 		foreach($recordIds as $recordId) {
 			if(Users_Privileges_Model::isPermitted($moduleName, 'Delete', $recordId)) {
+				// Inveteeで作成された子レコードは親が削除された際に削除されるので、改めて削除処理行われないようにする
+				$query = 'SELECT activityid 
+						  FROM vtiger_activity 
+						  WHERE invitee_parentid = ? 
+						  	AND deleted = 0 
+							AND activityid <> ? 
+						  ORDER BY activityid';
+				$result = $adb->pquery($query, array($recordId, $recordId));
+				for($i = 0; $i < $adb->num_rows($result); $i++) {
+					$skipRecords[$adb->query_result($result, $i, 'activityid')] = true;
+				}
+				
+				if (isset($skipRecords[$recordId])) {
+					unset($skipRecords[$recordId]);
+					continue;
+				}
+				
 				$recordModel = Vtiger_Record_Model::getInstanceById($recordId, $moduleModel);
 				$parentRecurringId = $recordModel->getParentRecurringRecord();
 				$adb->pquery('DELETE FROM vtiger_activity_recurring_info WHERE activityid=? AND recurrenceid=?', array($parentRecurringId, $recordId));
